Remove code used by frame_parallel_decode option.
Commit 5521a18fcd6c0248119f34a2584c174f183f4163 removed the
frame_parallel_decode option on the decoder side:
https://aomedia-review.googlesource.com/c/aom/+/50001

Remove more code used by the frame_parallel_decode option on the decoder
side.

Remove the AOM_CODEC_CAP_FRAME_THREADING flag.

Remove next_submit_worker_id, last_submit_worker_id, and
available_threads. This is the counterpart of
https://chromium-review.googlesource.com/c/webm/libvpx/+/557041.

Remove dthread.c and the related prototypes in dthread.h. This is the
counterpart of
https://chromium-review.googlesource.com/c/webm/libvpx/+/560621.

Remove the frame_worker_owner, 'row', and 'col' fields of RefCntBuffer.
This is the counterpart of
https://chromium-review.googlesource.com/c/webm/libvpx/+/560622.

Reduce FRAME_BUFFERS by 3. This is the counterpart of
https://chromium-review.googlesource.com/c/webm/libvpx/+/560620.

BUG=aomedia:2110

Change-Id: Iaec00020e5b2926507fea849fc3ab976042e1f81