Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Not a huge fan of this approach. It certainly works, but why not just switch to MSE for 4x4 blocks? At least for cdef_dist, it's supposed to be "comparable" to MSE.
> 
> (if the 4x4 vs 8x8 partition decision is hard, you could compute both in MSE, then use cdef for 8x8 to 16x16 compares and higher)

I agree that would be better (and that is in fact what this code used to do), but given the complexity of the codebase, it's a lot of work to get that working and keep it working. We also don't have good enough tests to keep people from breaking it (this would basically involve running MSE with and without dist-8x8 and ensuring we get the same results).

But this has been turned off at compile time since August, and is likely to bitrot further if it remains that way, so I think it is better to have something working (even if less than ideally) than to leave it broken and disabled.

Patch-set: 4
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>