Refine prune_comp_search_by_single_result
Use the best rd of GLOBALMV and NEWMV to prune the unlikely
reference frames for all the modes (NEARESTMV and NEARMV may not
have same motion vectors). Always keep the best of each mode
because it might form the best possible combination with other mode.

Tested BasketballDrill_832x480_50 (20 frames, bitrate=800, Speed 1),
it is 4% faster and coding performance 0.22% better.

STATS_CHANGED

Change-Id: I651308a78ef2ae501c288679847d7320783b3e3e