Update patch set 4
Patch Set 4:

WARNING: libaom__encoder_check/default detected a potential change:
If your commit message contains STATS_CHANGED this is just fyi. If it does not,
this is fatal.

In order to highlight changes which cause the encoder output to change,
you must update the commit message body with
STATS_CHANGED:
and a description of the expected behavior of the new encode process. For
example, if the motion search has been changed, or rate distortion weighting
changed, or a normative change to the bitstream has added or removed flags
or a feature.

bus_2_pass differs.

md5sum of webm output:
current:  6731c8e01b6338ea7e2bd3ac731d28a8
previous: 25947ef05b611608bfdc27c6594d7b59

file sizes:
31666 commit=current-low,config=bus_2_pass/file.webm
36569 commit=previous-low,config=bus_2_pass/file.webm

current  
previous 


To see the encode process look up the libaom__encoder_run which triggered this.
It is listed in the description here:
https://build.aomedia.org/jenkins/job/libaom__encoder_check/default/3773/

Patch-set: 4