Fix range check in half_btf() function
Per the linked bug report, the condition on 'result_64' in half_btf()
is not satisfied by all conformant bitstreams. But it can easily be
modified to give a condition which *is* true for all conformant bitstreams.

Add a comment explaining the updated condition, as well as its implications
for alternative implementations (ie. hardware and/or vectorized code)

BUG=aomedia:2151

Change-Id: Iab30344e2e7b1ce00245541b1a32a9495d5ed717