Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> Weird behavior. I'm always passing in a relative path (../libaom/build/cmake/toolchains/), it's just the second time through it throws a fit. Ah well, seems fine.

CMake converts the supplied relative path to an absolute path in the first run, but does not do so for the second and subsequent runs. The CMake behavior for the second and subsequent runs is the source of the problem. I'm not certain why CMake behaves this way, but it's not that big a deal as we're only using this to set the string in aom_config.c. The workaround is fairly simple-- just look at the CMAKE_TOOLCHAIN_FILE value and see if it starts with a '.'.

Patch-set: 3
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>