Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3: Code-Review+2
> > > 
> > > Weird behavior. I'm always passing in a relative path (../libaom/build/cmake/toolchains/), it's just the second time through it throws a fit. Ah well, seems fine.
> > 
> > CMake converts the supplied relative path to an absolute path in the first run, but does not do so for the second and subsequent runs. The CMake behavior for the second and subsequent runs is the source of the problem. I'm not certain why CMake behaves this way, but it's not that big a deal as we're only using this to set the string in aom_config.c. The workaround is fairly simple-- just look at the CMAKE_TOOLCHAIN_FILE value and see if it starts with a '.'.
> 
> Oh, I thought it was looking for '/' ... if it's looking  for '.' it could be misled but it would be a pretty weird directory configuration

I'm not really sure how it would be misled outside of someone intentionally trying to break the config check. Is there an example you have in mind? Keeping this check simple is nice, but if it needs to be expanded that's totally fine with me.

Note that looking for a leading '/' wouldn't work on Windows since the drive letter is the first part of an absolute path; would need a different set of tests/conditions.

Patch-set: 3