Add code that extract data for one pass svm.
This code fixes the eob calculation and estimates the eob and rdcost of
the four sub-blocks. Currently the estimations only works correctly when
skip flag is off.

Here, we save the EOB and rdcost in the struct RD_STATS when we do
transform. Then we pass the info in rdcost upstream until we reach
rd_pick_sb_modes, where to store the information in none_context. Due to
the complexity of the code, there might be some updates missing.

To turn on data generation, enable the CONFIG_ONE_PASS_SVM and
CONFIG_TRAIN_SVM flags in build/cmake/aom_config_defaults.cmake.

Change-Id: I626870fde623561f1943eea76a8b5cdc0e7a6714