VBR: Improve bit re-allocation per pyramid level.
Earlier, bits were reduced for leaf level frames and added for
internal ARF frames in the same loop. This meant that number of bits
reduced may NOT equal the number of bits added.

Now, we reduce bits from leaf level in the first loop, and then the
total of those bits is then reallocated to internal ARFs in a separate
loop. So, it's guaranteed that number of bits reduced = number of bits
added.

Overall BDRate (PSNR) improvement:

lowres: -0.594%
(garden_sif: -8.60%
 crew_cif:   -5.80%)

midres: -0.186%
(crew_4cif_30fps: -3.76%
 speed_bag_480p:  -1.45%)

STATS_CHANGED

Change-Id: I3cec7eaa980bc7a524e0a6ffdd50ee86327b7c04