Remove a check for an impossible condition.
The indexes stored in the cm->ref_frame_map array are either -1 or a
valid index in the range [0, FRAME_BUFFERS). So buf_idx >= FRAME_BUFFERS
cannot be true.

Add an assertion to verify that condition is not true. This assertion
may be removed after 2018-10-31, but it is fine to keep it.

BUG=aomedia:2179

Change-Id: I41b0e9ef53e76e275f6988d162636f6c90e9d4b2