Integerized SGR subspace projection
DO NOT MERGE: I have uploaded this patch to check against AWCY for any
significant compression performance change.

I have developed an integer-only implementation of the subspace
projection in the self-guided restoration filter encoder.  This is
mainly to serve as a more useful reference for hardware encoder
implementation.  This patch includes highbd and lowbd codepaths.

This change can be disabled by setting the build-time experiment flag
CONFIG_INTEGERIZE_SGR to 0

Change-Id: I98dc548dc17383a8bb42c5d00f0c06889c625d43