cfl_test: Fix rng initialization
61fc3e4c65 introduced a bug where the rnd declaration with
initialization as a local variable, like so:
ACMRandom rnd(0xBABA);

was replaced with splitting it as a class variable and supposedly
initializing it in the init() function, like so:
class {
 ACMRandom rnd;
 ...
 void init() {
   rnd(0xBABA);
 }
 ...
}

The second implementation is incorrect and instead of
initalization, it merely calls the overriden operator() in the
ACMRandom class. If that operator wasn't there in that class, this
line would've been an error.

Functionally, it doesn't change much since the rng is initiazed
with a deterministic seed by default. Getting of rid of this call
avoids a random number that is generated and thrown away.

Change-Id: I992c25e93392c5a5339ec1ba1866860701158879