Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I assume there is no case where this actually changes the decoder output (i.e. it is not a normative change), and this is just a tidy up (because when an unsigned value is read, the value will always be non-negative, and so the clamp will have no effect).
> 
> Do you agree?

Peter: Thank you for the review! You are right. When I wrote this CL, I was blindly comparing the libaom reference code with the spec. Later I also realized that the minimum of an unsigned value is 0, so this is just a tidy up, not a normative change.

Patch-set: 1
Reviewer: Gerrit User 9545 <9545@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>