Add support for CMAKE_INSTALL_LIBDIR and CMAKE_INS
TALL_INCLUDEDIR

Cmake will now use CMAKE_INSTALL_LIBDIR and CMAKE_INSTALL_INCLUDEDIR in
the installation path of includes and libraries. If not provided, it
will default to "include" and "lib".

${CMAKE_INSTALL_INCLUDEDIR} and ${CMAKE_INSTALL_LIBDIR} is now
passed on to pkg_config.cmake in order to generate a correct
aom.pc instead of using hardcoded directories.

BUG=aomedia:2195

Change-Id: I2561f92466920cea643664c037c8e4e97342aab