Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> FYI I can change it so there is no ABI change, but we'll have a dangling reference to the alpha channel bit

I see the idea is now to create an AV1 Alpha Image with mono_chrome set, with a AuxiliaryTypeProperty as specified in https://aomediacodec.github.io/av1-avif/#alpha-images - presumably this can be ignored by clients not using alpha.

But is changing AOM_IMG_FMT_HIGHBITDEPTH a good idea? What happens to existing files with high bit depths, or software reading them? Isn't there a risk that these may be incorrectly interpreted, unless implementers special-case v1-3?

Would it hurt to leave it at 0x800 but still bump the ABI? That way you could still make 0x400 available for other uses.

If there are no actual users of this standard right now this may be an academic concern, but I thought I'd mention it. In theory software should not try to load a file of a future version, but in practice if you juggle bits around I could totally see it trying to parse a 16-bit image as an 8-bit image with alpha (potentially exposing all kinds of bugs), especially if 16-bit images are rare.

Patch-set: 1
CC: Gerrit User 10287 <10287@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>