Refresh LAST_FRAME with error_resilient show_existing_frame
When error_resilient mode is disabled, the encoder side reference buffer
update for show_existing internal altref frames is handled as a special case,
where the ALTREF2_FRAME buffer is copied to the LAST_FRAME buffer. In
error_resilient mode, the show_existing_frame is re-encoded to avoid
producing an error_resilient frame with a dependency on previously
encoded frames. In this case, we still want to update LAST_FRAME, but
we want to do it from the current frame buffer rather than the
ALTREF2 buffer.

BUG=aomedia:2202, aomedia:2203

Change-Id: I99a7223ebbf145e1db0ae50a149584421f46c15f