Specify C++11 in build configuration
The build configuration currently does not specify a C++ standard to
target, relying on the compiler default.  Previously the codebase
compiled with C++03 but now C++11 features are in use breaking things
for older compilers.

(At the same time, change all instances of M_PI to the
internally-defined PI.  M_PI is non-standard and disappears from some
build environments with this change of C++ standard)

BUG=aomedia:2209

Change-Id: Ifb6c3dc1427e5148f93fd4727872ab1641d2ad29