Enforce C++11 in build configuration
The build configuration currently does not specify a C++ standard to
target, relying on the compiler default.  This might be C++03, C++11, or
C++14, depending on compiler and version.  We want to make sure
everybody is coding to the same standard.

MSVC doesn't support explicitly setting the standard to C++11.  But
newish versions default to C++14 which should compile C++11 code, and
the build tests will prevent MSVC users accidentally submitting C++14
code.

Also, change all instances of M_PI to the internally-defined PI.  M_PI
is non-standard and disappears from some build environments with this
change of C++ standard.  Also remove the target-specific C++11 flags
used for dump_obc and libwebm.

BUG=aomedia:2209

Change-Id: Ifb6c3dc1427e5148f93fd4727872ab1641d2ad29