av1_highbd_dr_prediction_z2_c: Add an assertion.
We add an assertion to make sure that we are not reading beyond the valid range
of the 'above' or 'left' buffers.

Valid minimum index is:
-1 when upsampling is NOT used, and
-2 when upsampling is used.
And this minimum index corresponds to the top-left pixel.

Also,
- A similar assertion was already present in lowbd function
av1_dr_prediction_z2_c. But I made the function structure same as hbd
version.
- Related code cleanup for variable names, and moving declarations
closer to use.

Change-Id: If477013843edcf408332b107e58a2238ca30dfb7