Update patch set 30
Patch Set 30:

> Patch Set 30:
> 
> > Patch Set 30:
> > 
> > Thanks for the patch.
> > 
> > Discussed with Scott, if it is possible, could you reduce the code size? Also, limit the using of AOM_FORCE_INLINE as we suggested before.
> 
> Hi, All. Thanks a lot for your review. It is the very first vectorized version of directional prediction, to  be improved further for sure. As for the code size - yes, I plan to decrease it later.
>  As for the AOM_FORCE_INLINE - I'm sure it is located where it is absolutely necessary only. Just a friendly reminder - the situation when a large function is surrounded by the very tiny pieces of code is not inlined by existing compilers even if inline is specified. And it leads to significant performance drop unfortunately :( But I will double check my code for that.
> As for the timing data - It has been copy paste by Eirk from the very old patch I've used as a starting point for my own one. I will re-run it and change the commit message accordingly.

Please use the speed test in dr_prediction_test.cc.  The speed test was added in June mostly to check the performance of a SSE4 LBD version dr predictor done in mid May and to verify the AVX2 HBD performance numbers.  The SSE4 version was put on hold because the performance was not as good as hoped, but at the time, comparable to the AVX2 version.  The SSE4 version also had ASAN errors, so correcting those would probably further decrease performance.

Patch-set: 30