Update patch set 5
Patch Set 5:

Hi Yunqing,
We tested the patch for wikipedia 1080p content @ 1mbps and 3 mbps. On average, we see 6.7%, -0.5% and .1% encode time reduction for 4,5 and 8 presets.
We tried disabling 3 speed features for higher presets as well (as an experiment) and observed that encode time reduction is -4%, 6.4%,-1.1% for 1,2 and 3 presets.
Overall we can see that speed data is encouraging except speed=1 or 3 presets. Hence we feel that for speed >= 4, the patch gives a good quality improvement with sizable reduction or negligible increase in encode time.
We can revisit the hash speed features for speed=1,2 or 3 separately (by experimenting with each of the hash speed features).
Please let us know your thoughts.

Patch-set: 5