Form CurrentFrame structure, rename frame_offset
Move a number of elements of AV1_COMMON which refer specifically to the
current frame into a new CurrentFrame structure.  Also, move those
elements specific to skip mode into a SkipModeInfo structure.  This
change is not exhaustive; we plan to move over more elements in future.

Also, frame_offset is renamed to order_hint in order to be clearer about
its purpose and to match the terminology in the bitstream spec.

Change-Id: Ie85280138d11b83625a764602bef502252a4567f