Update patch set 31
Patch Set 31:

> Patch Set 30: Code-Review-2
> 

> I reverted the changes to the test locally, and seems to work fine.
> 
> To prevent this from accidentally being submitted, I'm going to block.

Hi, Scott. I've changed dr_prediction_test.cc to make the Speedtest  pass for lbd as well.
However please notice that currently running it doesn't make sense for lbd or for a pure C version :)
So could you please do me a favor and +2 this patch as a first AVX2 implementation version. 
My second patch featuring Z2 code and some of Z1-Z3 improvements is 100% ready, but it is based on this patch..

Patch-set: 31