Do not acquire extra fb references.
The cm->ref_frame_map and cm->next_ref_frame_map arrays hold references
to frame buffers. The policy is that whenever we store a frame buffer
index in cm->ref_frame_map or cm->next_ref_frame_map, we increase the
ref_count of that frame buffer. Conversely, whenever we overwrite an
element of these arrays with another value, we need to decrease the
ref_count of the old frame buffer.

Under this policy, generate_next_ref_frame_map() acquires an extra frame
buffer reference in the "if (mask & 1)" case. This still works out OK
because both release_frame_buffers() and swap_frame_buffers() compensate
for the extra extra frame buffer references, but it makes the code very
confusing.

This CL fixes generate_next_ref_frame_map() so that it does not acquire
the extra frame buffer references, and makes the corresponding changes
to release_frame_buffers() and swap_frame_buffers().

BUG=aomedia:2266

Change-Id: I88fb9c06f17fb5d3ec77ca40c063f763af88e9de