Fix/cleanup encoder frame_id_numbers_present_flag
frame_id_numbers_present_flag was being set in a variety of places to
conflicting values.  Furthermore, a combination of these meant that the
encoder could never actually produce a bitstream with this flag enabled.

I have cleaned this up so the flag is set only once.  The flag is
enabled if the error_resilient_mode setting is enabled in the
encoder (unless large-scale-tile or reduced still picture headers are
enabled, in which case the flag is disabled).

Using constant-quality mode and error-resilient mode, this patch
modifies the output bitstream but the bitstream decode result and PSNR
are unchanged.  Using constant-bitrate mode and error-resilient mode,
the decode output is changed (a slight PSNR reduction is to be
expected).  With error resilient mode disabled, the bitstream is
unchanged.

STATS_CHANGED

Change-Id: I5ad2a1317fec5b15bc7b2bb5a9b49f2f3123e931