Fix/cleanup encoder frame_id_numbers_present_flag
frame_id_numbers_present_flag was being set in a variety of places to
conflicting values.  Furthermore, a combination of these meant that the
encoder could never actually produce a bitstream with this flag enabled.

I have cleaned this up so the flag is set only once.  The flag is
enabled if the error_resilient_mode setting is enabled in the
encoder (unless large-scale-tile or reduced still picture headers are
enabled, in which case the flag is disabled).

Using constant-quality and error-resilient mode, this patch increases
the output bitstream size but the decode result and PSNR are unchanged.
Using constant-bitrate mode and error-resilient mode, the decode output
is changed (a slight PSNR reduction is to be expected).  With error
resilient mode disabled, the bitstream is unchanged with this patch.
If the increased size in error-resilient mode is undesirable then
frame_id_numbers_present_flag could be controlled separately.
https://arewecompressedyet.com/?job=base%402018-11-28T15%3A04%3A43.665Z&job=test%402018-11-28T18%3A13%3A36.897Z

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder.
BUG=aomedia:2244

STATS_CHANGED

Change-Id: I5ad2a1317fec5b15bc7b2bb5a9b49f2f3123e931