Adjustments to superes qthresh mode
Obtain the superres parameters directly for the q being used.
Also the kf qthresh default is made a low 32 by default since
the analysis already does a decent job deciding whether to use
superres and by how much.

This effectively reduces the dependence on the qthresh parameter
and may eventually be removed.

Change-Id: Ie1d5204bc029615222d612f3a671bf12fc767f57