Update patch set 3
Patch Set 3:

> Hi, Chang, thanks again for your efforts. The stride could be guaranteed to be a multiple of 8. (but not of 16!)

Hi Victoria,

In transpose_TX_8X8(), the stride is added to dst, which is a uint16_t* pointer:

static void transpose_TX_8X8(const uint16_t *src, ptrdiff_t pitchSrc,
                             uint16_t *dst, ptrdiff_t pitchDst) {
  ...
  _mm_storeu_si128((__m128i *)(dst + 0 * pitchDst), r0_Lo);
  _mm_storeu_si128((__m128i *)(dst + 1 * pitchDst), r2_Lo);
  _mm_storeu_si128((__m128i *)(dst + 2 * pitchDst), r1_Lo);
  _mm_storeu_si128((__m128i *)(dst + 3 * pitchDst), r3_Lo);
  ...
  _mm_storeu_si128((__m128i *)(dst + 4 * pitchDst), r0);
  _mm_storeu_si128((__m128i *)(dst + 5 * pitchDst), r2);
  _mm_storeu_si128((__m128i *)(dst + 6 * pitchDst), r1);
  _mm_storeu_si128((__m128i *)(dst + 7 * pitchDst), r3);
}

Therefore, if the stride is a multiple of 8, then the resulting pointer will be 16-byte aligned. So we should be able to use the aligned _mm_store_si128() function instead.

Patch-set: 3