Fix inter-frame film grain params
When using film grain for a shown inter-frame, there is the option to
re-use the parameters from a reference frame.  Logically, the encoder
should indicate a reference frame (using film_grain_params_ref_idx) with
the same film grain parameters as the current frame.  However, the
encoder appears to do the opposite, possibly because of a
misunderstanding of the result of memcmp.  This patch resolves this so
that the encoder actually selects a reference frame with matching
parameters.

Change-Id: I3d97213ab8b2d77a0ac5b842100785590800a01c