Fix inter-frame film grain params
When using film grain for a shown inter-frame, there is the option to
re-use the parameters from a reference frame.  Logically, the encoder
should indicate a reference frame (using film_grain_params_ref_idx) with
the same film grain parameters as the current frame.

Previously, the encoder had a bug where it would actually choose a
reference frame where the film grain parameters did *not* match.
Furthermore, a simple use of memcmp to compare the film grain parameters
has two issues: structure packing values do not have a guaranteed value
unless the structure is initialised using memset, and two attributes of
the structure are overwritten and so their equality does not matter
(update_parameters and random_seed).

This patch replaces the use of memcmp with an equivalence checking
function for film grain parameter structures.

Change-Id: I3d97213ab8b2d77a0ac5b842100785590800a01c