Remove frame_refs: de-duplicate ref frame refs
Currently, there are multiple, duplicate, references to reference
frames.  This patch removes one of those, the frame_refs array, using
remapped_ref_idx instead.  remapped_ref_idx is moved from AV1_COMP to
AV1_COMMON so it can be used in the same way in both the encoder and
decoder.  Functions are added for convenient access to reference frame
buffers.

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder.

BUG=aomedia:2244

Change-Id: Ib7c65abecbdd7a9d64dfea4a2d5a173327253812