Make tile_buffers local, fix LST endianness bug
I move tile_buffers[] from AV1_COMP to being a local in
bitstream.c:write_tiles_in_tg_obus() as it is not used elsewhere.  This
caused the following bug to surface:

To indicate a LST copy-tile, write_tiles_in_tg_obus() sets the most
significant bit of the tile size uint32_t.  Later, attempting to read
this bit, find_largest_tile() reads the most significant bit in the
first byte of the header.  However, on a little-endian system the MSbit
will be in the last byte, not the first.

This bug seemed to be causing reads outside tile_buffers[] in the
lightfield test.  This went unnoticed while tile_buffers[] was inside
cpi on the heap, but caused seg-faults when tile_buffers[] was moved to
the stack.  Also, this may have caused erroneous output in LST TGs.

Change-Id: I3cb76487ed4ba5591ef77b628e074264628f07b6