Skip affine evaluation based on warp parameters
Introduced a speed feature to gate wrap function evaluation for motions of type IDENTITY,
TRANSLATION and AFFINE based on number of warp neighbors.
This speed feature is turned on for speeds >= 3.

As per AWCY runs for speed=3 and 4, the following are the observed BD-rate & encoder time reduction numbers.

Preset  BD-Rate-Drop/Gain   Enc-time-Reduction
  4         -0.02               ~0.85 (averaged across multiple test cases)
  3          0.01               ~0.65 (averaged across multiple test cases)

STATS_CHANGED

Change-Id: Ic7e11408023578602244b165bb34ab4cbfff28a7